Tab transitions were removed for the five main screens.  Each screen now loads immediately. PlantInfoCard edit of Repot and Division fields showed Input not Date Picker.  This resulted in an error for the plant streams (string not int).  Fixed. Growth Cycle? Additional layers to guarantee data: Additional try/catch loops. Additional null checks for incoming data. Data type check for every piece of data incoming from the database. Data type checks on variable input data to prevent wrong data entry. DialogItem for DialogSelect now registers even onTap of transparent areas. Wrapped the entire bloom sequence in edit onTap instead of just the edit button. Creation of Data Validation Class with incoming type checks for boolean and number: Extended to PlantData, now have data type checks for string, int, double, bool, and list. PlantTile plant name container width has been increased allowing more text to be displayed. Message documents now include participants list. Addition of dormancy cycle and bug fixes for bloom sequence. Recent updates: Addition of update bubble on plant tile. Addition of recent updates bades to connectionCard, searchTileUser, and profile header. UserData updates: Date of lastPlantAdded. Date of lastPlantUpdated. Show wishlist. Show sell list. Top Collectors no longer takes a cached file, it queries top users. Revamp of images: Creation of ImageData which houses thumbnail, full image, and date. Addition of List of ImageData to PlantData. Revision of all image creation/deletion functions. Resolved widget change thumbnail load issue when switching from carousel to gridview. PlantCount now excludes plants in import, clone, or wishlists. Wishlist and Sell List implementation: Denoted with a star Create sets as wanted true. Move from another shelf is now allowed Move to another shelf set the value for wanted/sell as false in dialog_functions for Collection. Create shelves on library load. Account settings added to enable/disable. Download Links: Addition of sharelinks as constants with appropriate references. Share plant provides links and basic data Share if email not found in add by email. Share app button added in Friends.

+++ functions/src/index.ts
@@ -3,42 +3,27 @@ import * as admin from "firebase-admin";
 const firebase = admin.initializeApp();
 
 export const deletePlantPhotos = functions.firestore
-  .document("plants/{plantId}")
-  .onDelete((snap, context) => {
-    const { userId } = context.params;
-    const { plantId } = context.params;
-    const bucket = firebase.storage().bucket();
+    .document("plants/{plantId}")
+    .onDelete((snap) => {
+        if (snap) {
+            const data = snap.data();
+            if (data) {
+                const userId = data.plantOwner;
+                const plantId = data.plantID;
+                const bucket = firebase.storage().bucket();
 
-    return bucket.deleteFiles({
-      prefix: `users/${userId}/plants/${plantId}`
+                return bucket.deleteFiles({
+                    prefix: `users/${userId}/plants/${plantId}`
+                });
+            } else {
+                console.log('No data in snapshot');
+                return;
+            }
+        } else {
+            console.log('No snapshot information');
+            return;
+        }
     });
-  });
-
-export const getTopCollectors = functions.firestore
-.document("users/{userId}")
-.onWrite(async () => {
-//***ON WRITE TO USER FILE, GET AND CACHE TOP RANKED USERS
-
-  //get the data
-  const snapUsers = await admin.firestore().collection("users/")
-  .orderBy("userTotalPlants", "desc").limit(50).get();
-  //use .docs() to get an array of DocumentSnapshots for QuerySnapshot
-  //but just use forEach to skip this first step
-  const data: any[] = [];
-  snapUsers.forEach(userSnap => {
-      const object = userSnap.data();
-      data.push(object);
-  });
-
-  //if no, nothing can be done
-  if (data === []) return null;
-
-  //remove the old value and user reference
-  return admin.firestore().doc("app/user_stats_top").update({
-        byPlants : data
-        });
-//FUNCTION END
-});
 
 export const updateUserFriendList = functions.firestore
 .document("users/{userID}/userConnections/{friend}")
